% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.declaration.R
\name{create.declaration}
\alias{create.declaration}
\title{Creates variable declarations for Stan programs}
\usage{
create.declaration(var.name, var.type, var.dim)
}
\arguments{
\item{var.name}{String indicating the name of the variable}

\item{var.type}{String indicating the variable type and any restrictions}

\item{var.dim}{String or number indicating the dimension of the variable}
}
\value{
A string for the resulting variable declaration.
}
\description{
Creates variable declarations for Stan programs
}
\details{
This method allows for easy creation of variable declarations, accounting for semantic
differences in declaring vectors vs. arrays.

If var.dim is a string, it should be in the form of dimension declarations in a Stan program
(e.g., "[N, k]").
}
\examples{
create.declaration("foo", "real<lower=0>", 5)
create.declaration("bar", "vector", "[k]")
create.declaration("mat", "matrix", "[N, k]")

}
