% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.restricted.models.R
\name{create.restricted.models}
\alias{create.restricted.models}
\title{Creates restricted models for IKDE}
\usage{
create.restricted.models(ikde.model, eval.point)
}
\arguments{
\item{ikde.model}{An object of class ikde.model, does not necessarily have to be built}
}
\value{
Returns a list of built ikde.models for each restricted model
}
\description{
Creates set of restricted models to be used for posterior density estimation
}
\details{

}
\examples{
data(lm.generated)

X <- lm.generated$X
y <- lm.generated$y

data <- list(N = list("int<lower=1>", nrow(X)),
             k = list("int<lower=1>", ncol(X)),
             X = list("matrix[N, k]", X),
             y = list("vector[N]", y))
parameters <- list(beta = "vector[k]",
                   sigma = "real<lower=0>")
model <- list(priors = c("beta ~ normal(0, 10)",
                         "sigma ~ inv_gamma(1, 1)"),
              likelihood = c("y ~ normal(X * beta, sigma)"))

ikde.model <- define.model(data, parameters, model)
eval.point <- list(beta = c(0, 1, 2, 3),
                   sigma = 1)

parameter <- names(ikde.model$parameters)[1]
vector.index <- 1

}
