% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.expression.R
\name{evaluate.expression}
\alias{evaluate.expression}
\title{Evaluate expression from Stan program}
\usage{
evaluate.expression(stan.expression)
}
\arguments{
\item{stan.expression}{String representing Stan expression. All variables must be present in parent environment.}
}
\value{
The result of the Stan expression
}
\description{
Evaluate expression from Stan program
}
\details{

}
\examples{
X <- matrix(1:9, nrow = 3)
b <- c(4, 5, 6)

stan.expression <- "(3 + 2) * X * (5 * b)"

# These results match:
evaluate.expression(stan.expression)
print((3 + 2) * X \%*\% (5 * b))

}
