% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.model.R
\name{fit.model}
\alias{fit.model}
\title{Define Stan model}
\usage{
fit.model(ikde.model, burn.iter = 1000, sample.iter = 1000,
  chains = 1, control = NULL, refresh = NULL,
  display.output = FALSE)
}
\arguments{
\item{ikde.model}{An object of class ikde.model which has been built}
}
\value{
An object of S4 class stanfit. See rstan::stan for more details.
}
\description{
Defines Stan model, creates model code, and stores input data
}
\details{

}
\examples{
  
}
\references{

}
