% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm.generated}
\alias{lm.generated}
\title{Randomly generated multivariate linear model data}
\format{A list with two components:
\describe{
  \item{X}{Matrix of independent variables}
  \item{y}{Vector of dependent variable observations}
}}
\usage{
lm.generated
}
\description{
A dataset for estimation of linear models
}
\keyword{datasets}
