% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified.exp.lower.inv.R
\name{modified.exp.lower.inv}
\alias{modified.exp.lower.inv}
\title{Modified exponential function with lower bound}
\usage{
modified.exp.lower.inv(x, lower.limit = 0)
}
\arguments{
\item{x}{Point to evaluate modified exponential function and gradient}

\item{lower.limit}{Lower limit of the modified exponential function; defaults to 0}
}
\value{
Returns a list with two elements, each with the same number of elements as x
\item{value}{Value of the inverse modified exponential function at x}
\item{gradient}{Value of the gradient of the inverse modified exponential function at x}
}
\description{
Modified exponential function with lower bound
}
\details{
Evaluates the inverse modified exponential function (log(x - lower.bound)) and its gradient (1 / (x - lower.bound)).
}
\examples{
x <- seq(0.005, 5, length.out = 100)
exp.result <- modified.exp.lower.inv(x)

plot(exp.result$value ~ x, type = "l")
plot(exp.result$gradient ~ x, type = "l")

}
