% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified.exp.lower.R
\name{modified.exp.lower}
\alias{modified.exp.lower}
\title{Modified exponential function with lower bound}
\usage{
modified.exp.lower(x, lower.limit = 0)
}
\arguments{
\item{x}{Point to evaluate modified exponential function and gradient}

\item{lower.limit}{Lower limit of the modified exponential function; defaults to 0}
}
\value{
Returns a list with two elements, each with the same number of elements as x
\item{value}{Value of the modified exponential function at x}
\item{gradient}{Value of the gradient of the modified exponential function at x}
}
\description{
Modified exponential function with lower bound
}
\details{
Evaluates the modified exponential function (exp(x) + lower.bound) and its gradient (exp(x)).
}
\examples{
x <- seq(-5, 5, length.out = 100)
exp.result <- modified.exp.lower(x)

plot(exp.result$value ~ x, type = "l")
plot(exp.result$gradient ~ x, type = "l")

}
