% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified.exp.upper.R
\name{modified.exp.upper}
\alias{modified.exp.upper}
\title{Modified exponential function with upper bound}
\usage{
modified.exp.upper(x, upper.limit = 0)
}
\arguments{
\item{x}{Point to evaluate modified exponential function and gradient}

\item{upper.limit}{Lower limit of the modified exponential function; defaults to 0}
}
\value{
Returns a list with two elements, each with the same number of elements as x
\item{value}{Value of the modified exponential function at x}
\item{gradient}{Value of the gradient of the modified exponential function at x}
}
\description{
Modified exponential function with upper bound
}
\details{
Evaluates the modified exponential function (-exp(x) + upper.bound) and its gradient (-exp(x)).
}
\examples{
x <- seq(-5, 5, length.out = 100)
exp.result <- modified.exp.upper(x)

plot(exp.result$value ~ x, type = "l")
plot(exp.result$gradient ~ x, type = "l")

}
