% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.inv.R
\name{sigmoid.inv}
\alias{sigmoid.inv}
\title{Inverse sigmoid function and gradient}
\usage{
sigmoid.inv(x, lower.limit = 0, upper.limit = 1)
}
\arguments{
\item{x}{Point to evaluate inverse sigmoid function and gradient}

\item{lower.limit}{Lower limit of the sigmoid function; defaults to 0}

\item{upper.limit}{Upper limit of the sigmoid function; defaults to 1}
}
\value{
Returns a list with two elements, each with the same number of elements as x
\item{value}{Value of the inverse sigmoid function at x}
\item{gradient}{Value of the gradient of the inverse sigmoid function at x}
}
\description{
Inverse sigmoid function and gradient
}
\details{
Evaluates the inverse sigmoid function (-log((upper.bound - lower.bound) / (x - lower.bound) - 1)) and its gradient ((upper.bound - lower.bound) / ((lower.bound - x) * (x - upper.bound))).
}
\examples{
x <- seq(0.005, 0.995, length.out = 100)
sigmoid.inv.result <- sigmoid.inv(x)

plot(sigmoid.inv.result$value ~ x, type = "l")
plot(sigmoid.inv.result$gradient ~ x, type = "l")

}
