% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.multiply.R
\name{\%stan*\%}
\alias{\%stan*\%}
\title{Function to replicate multiplication in Stan}
\usage{
x \%stan*\% y
}
\arguments{
\item{x}{First term in product}

\item{y}{Second term in product}
}
\value{
Returns an object of the same type as the base
}
\description{
Function to replicate multiplication in Stan
}
\examples{

}
