% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.model.R
\name{define.model}
\alias{define.model}
\title{Define Stan model}
\usage{
define.model(data, parameters, model, transformed.data = list(),
  transformed.parameters = list())
}
\arguments{
\item{data}{A list of data passed to the Stan program. Should be of the form list(data.name = list(data.type, data.object)).}

\item{parameters}{A list of parameters used in the Stan program. Should be of the form list(parameter.name = parameter.type).}

\item{model}{A list describing the Stan model. Should be a list with components "priors" and "likelihood".}

\item{transformed.data}{A list describing data transformations for the Stan program to perform. Should be of the form list(variable.name = list(variable.type, variable.expression)).}

\item{transformed.parameters}{A list describing parameter transformations for the Stan program to perform. Should be of the form list(variable.name = list(variable.type, variable.expression)).}
}
\value{
Returns a list with the following elements
\item{y.fit}{Estimated value of the frontier at X.fit}
}
\description{
Defines Stan model, creates model code, and stores input data
}
\details{

}
\examples{
  
}
\references{

}
