% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.priors.R
\name{evaluate.priors}
\alias{evaluate.priors}
\title{Define Stan model}
\usage{
evaluate.priors(ikde.model, eval.point)
}
\arguments{
\item{ikde.model}{An object of class ikde.model which has been built}

\item{eval.point}{A list of parameter names and the point to evaluate priors}
}
\value{
A real number indicating value of the log-prior at the evaluation point
}
\description{
Defines Stan model, creates model code, and stores input data
}
\details{
Parses sampling statements in ikde.model$model$priors and evaluates them at the specified
evaluation point.
}
\examples{
data(lm.generated)

X <- lm.generated$X
y <- lm.generated$y

data <- list(N = list("int<lower=1>", nrow(X)),
             k = list("int<lower=1>", ncol(X)),
             X = list("matrix[N, k]", X),
             y = list("vector[N]", y))
parameters <- list(beta = "vector[k]",
                   sigma = "real<lower=0>")
model <- list(priors = c("beta ~ normal(0, 10)",
                         "sigma ~ inv_gamma(1, 1)"),
              likelihood = c("y ~ normal(X * beta, sigma)"))

ikde.model <- define.model(data, parameters, model)
ikde.model <- build.model(ikde.model)
stan.fit <- fit.model(ikde.model)
stan.extract <- extract(stan.fit)

print(apply(stan.extract$beta, 2, mean))
  
}
