% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R, R/sigmoid.inv.R, R/stan.multiply.R
\name{sigmoid}
\alias{sigmoid}
\alias{sigmoid.inv}
\alias{\%stan*\%}
\title{Sigmoid function and gradient}
\usage{
sigmoid(x)

sigmoid.inv(x)

x \%stan*\% y
}
\arguments{
\item{x}{Point to evaluate sigmoid function and gradient}

\item{y}{Second term in product}

\item{x}{Point to evaluate inverse sigmoid function and gradient}

\item{x}{First term in product}
}
\value{
Returns a list with two elements, each with the same number of elements as x
\item{value}{Value of the sigmoid function at x}
\item{gradient}{Value of the gradient of the sigmoid function at x}

Returns a list with two elements, each with the same number of elements as x
\item{value}{Value of the sigmoid function at x}
\item{gradient}{Value of the gradient of the sigmoid function at x}

Returns an object of the same type as the base
}
\description{
Sigmoid function and gradient

Inverse sigmoid function and gradient

Function to replicate multiplication in Stan
}
\details{
Evaluates the sigmoid function (1 / (1 + exp(-x))) and its gradient (exp(-x) / (1 + exp(-x))^2).

Evaluates the inverse sigmoid function (-log(1 / x - 1)) and its gradient (1 / (x - x^2)).

Accepts arguments x and y. If either is a singleton, returns the value of x*y (in R notation).
If both arguments are matrices or vectors, returns x%*%y (in R notation).
}
\examples{
x <- seq(-5, 5, length.out = 100)
sigmoid.result <- sigmoid(x)

plot(sigmoid.result$value ~ x, type = "l")
plot(sigmoid.result$gradient ~ x, type = "l")

x <- seq(0.005, 0.995, length.out = 100)
sigmoid.inv.result <- sigmoid.inv(x)

plot(sigmoid.inv.result$value ~ x, type = "l")
plot(sigmoid.inv.result$gradient ~ x, type = "l")

X <- matrix(1:9, nrow = 3)
b <- c(4, 5, 6)

(3 + 2) * X \%stan*\% (5 * b)
#      [,1]
# [1,] 1650
# [2,] 2025
# [3,] 2400

}
